food_ai/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=3.0.0
transformers>=4.30.0
gradio>=4.0.0
gymnasium>=0.29.0
numpy>=1.24.0
Pillow>=9.0.0
